rename class and remove try catch
