+++ Bookish/CSVReader.cs
@@ -1,0 +1,37 @@
+﻿using Bookish.DataAccess;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Bookish.ConsoleApp
+{
+    class CSVReader
+    {
+        public static List<string> ExtractFromCSV(string path, bool includesHeaderRow)
+        {
+            int linesToSkip = includesHeaderRow ? 1 : 0;
+
+            string[] text = File.ReadAllLines(path);
+            List<string> allLines = text.Skip(linesToSkip).ToList();
+            return allLines;
+        }
+
+        public static void AddCSVLineToBookDatabase(List<string> allLines)
+        {
+            foreach (string line in allLines)
+            {
+                List<string> item = line.Split(",").ToList();
+
+                string title = item[0].Trim();
+                string author = item[1].Trim();
+                string isbn = item[2].Trim();
+                string coverPhotoUrl = item.Count < 4 ? "https://islandpress.org/sites/default/files/default_book_cover_2015.jpg" : item[3].Trim();
+
+                BookRepo.AddBook(title, author, isbn, coverPhotoUrl);
+            }
+        }
+    }
+}
